Remove dependency to mj-include component
